Crate example_test

Macros

run_current_example

Runs the current example’s main in a child process, returning an ExampleChild handle. Only arguments need to be specified.

Structs

ExampleChild

A wrapper around std::process::Child that allows us to wait for a specific outputs.

Constants

RUN_MAIN_ENV_VAR

Environment variable used to signal that the current test binary should run the example’s main instead of the test harness. Set to "1" by ExampleChild::run_new.

Traits

ExampleMainReturn

Return types allowed for an example main run via run_current_example!.

Functions

run_example_main_if_child

Used by run_current_example!; do not call directly.