example_test

Trait ExampleMainReturn

pub trait ExampleMainReturn {
    // Required method
    fn into_exit_code(self) -> i32;
}

Return types allowed for an example main run via run_current_example!.

Required Methods

fn into_exit_code(self) -> i32

Converts the return value of main into a process exit code.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".