Crate example_test

Macros

run_current_example

Calls ExampleChild::run_new with the package name and test name infered from the call site. Only arguments need to be specified.

Structs

ExampleChild

A wrapper around std::process::Child that allows us to wait for a specific outputs.

Functions

extract_example_name

Extract the example name from the std::file! path, used by the run_current_example! macro.